Add conversion operators between vectors
Vectors should have the ability to convert to and from other vectors of
the same dimension, where only the value_type differs.  A conversion
operator that allows vectors to convert to the next dimension up is
provided, the last vector component is assigned a default value of 1
(see homogenous coordinates).